refactor(engine): S2 Stage 6 — extract EQ/Waveshaper/SmartBalance bindings to XlethSvcEqWs TU
verbatim move of 15 handlers + 3 in-block static helpers (getEQ/getWS/getSmartBalance, no promotion); zero behavior change; ctest matches baseline.
